Fixed bug causing incorrect navigate behavior.
If navigate was set to true and a navigation behavior was inferred from a hyperlink, after which the clipboard instance was transferred to a non-hyperlink or a hyperlink with a javascript URL, the navigation URL was not cleared.

This caused "sticky" navigation behavior unless explicity cleared, which defeats the purpose of the separate "navgiate" boolean option altogether.